% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{get_power}
\alias{get_power}
\title{Get result table of the fitted model}
\usage{
get_power(expt, d, p = 0, nsim = 100, seed = 2018)
}
\arguments{
\item{expt}{An object of class "expt"}

\item{d}{A vector of effect sizes (difference in mean / s.d.)}

\item{p}{Intraclass correlation coefficient(proportion of variance
explained by cage), between 0 to 1}

\item{nsim}{Number of simulations per value of d}

\item{seed}{Random seed}
}
\value{
a tibble with estimated power corresponding to d, p and standard t test power
calculated based on the minim number of mice per cage if you have unqual cages.
}
\description{
Get result table of the fitted model
}
