% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{make_expt}
\alias{make_expt}
\title{Make an experiment object}
\usage{
make_expt(
  ncage_treatment = 4,
  ncage_control = ncage_treatment,
  mice_per_cage_treatment = 1,
  mice_per_cage_control = mice_per_cage_treatment
)
}
\arguments{
\item{ncage_treatment}{The number of cages per treatment, in default it equals to the control.}

\item{mice_per_cage_treatment}{number of mice in each cage in treatment group, in default it equals to the control.}
}
\value{
An object of class "expt"
}
\description{
Make an experiment object
}
